norm

+++ srcs/power.c
@@ -6,7 +6,7 @@
 /*   By: mery <mery@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/05/15 14:51:02 by jmery             #+#    #+#             */
-/*   Updated: 2020/09/23 12:37:22 by mery             ###   ########.fr       */
+/*   Updated: 2020/09/29 13:19:45 by mery             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,10 +71,10 @@ void	clean_power(char *value)
 	}
 }
 
-size_t	get_power_index_last(char *value)
+int		get_power_index_last(char *value)
 {
 	int		index;
-	size_t	i;
+	int		i;
 
 	index = -1;
 	i = ft_strlen(value) - 1;
